==> STEP_0_PACKAGES_TO_INSTALL.R <==

#General Use
install.packages("shiny")
install.packages("shinydashboard")
install.packages("shinythemes")
install.packages("rmarkdown")
install.packages("rmdformats")
install.packages("httr")
install.packages("jsonlite")
install.packages("dplyr")
install.packages("lubridate")
install.packages("plotly")
install.packages("kableExtra")

# NLP
install.packages("purrr")
install.packages("tidytext")
install.packages("topicmodels")
install.packages("reshape2")
install.packages("irlba")
install.packages("ggraph")
install.packages("igraph")
install.packages("networkD3")
install.packages("tm")


#Geographic
install.packages("plyr")
install.packages("tidyr")
install.packages("rgdal")
install.packages("sf")
install.packages("leaflet")
