==> Step_9_Null_Report.Rmd <==
---
title: "Report Not Rendered"
author: "APD - TPCHD"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)
```
## Dates Provided / Syndrome did not have sufficient visits (>1) to generate report.
